@@ -1,11 +1,13 @@
 import 'package:firebase_auth/firebase_auth.dart';
 
-class FirebaseHelper {
-
+class FirebaseServices {
   final _auth = FirebaseAuth.instance;
 
   signIn(String email, String password) async {
-      return _auth.signInWithEmailAndPassword(email: email, password: password);
+    return _auth.signInWithEmailAndPassword(email: email, password: password);
   }
 
+  forgotPassword(email) async {
+    return _auth.sendPasswordResetEmail(email: email);
+  }
 }
